refactor: move RetryConfigurationResolver to runtime and use it
