+++ runtime/src/main/java/io/quarkiverse/operatorsdk/runtime/ConfigurationServiceRecorder.java
@@ -5,7 +5,6 @@ import java.util.function.Supplier;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.javaoperatorsdk.operator.api.config.ControllerConfiguration;
-import io.javaoperatorsdk.operator.api.config.RetryConfiguration;
 import io.javaoperatorsdk.operator.api.config.Version;
 import io.quarkus.arc.Arc;
 import io.quarkus.runtime.annotations.Recorder;
@@ -31,21 +30,7 @@ public class ConfigurationServiceRecorder {
             if (extConfig != null) {
                 extConfig.finalizer.ifPresent(c::setFinalizer);
                 extConfig.namespaces.ifPresent(c::setNamespaces);
-                extConfig.retry.ifPresent(r -> {
-                    long initialInterval = RetryConfiguration.DEFAULT_INITIAL_INTERVAL;
-                    double intervalMultiplier = RetryConfiguration.DEFAULT_MULTIPLIER;
-                    long maxInterval = RetryConfiguration.DEFAULT.getMaxInterval();
-                    if (r.interval.isPresent()) {
-                        final var intervalConfiguration = r.interval.get();
-                        initialInterval = intervalConfiguration.initial.orElse(RetryConfiguration.DEFAULT_INITIAL_INTERVAL);
-                        intervalMultiplier = intervalConfiguration.multiplier.orElse(RetryConfiguration.DEFAULT_MULTIPLIER);
-                        maxInterval = intervalConfiguration.max.orElse(RetryConfiguration.DEFAULT.getMaxInterval());
-                    }
-                    final var retry = new PlainRetryConfiguration(
-                            r.maxAttempts.orElse(RetryConfiguration.DEFAULT_MAX_ATTEMPTS), initialInterval,
-                            intervalMultiplier, maxInterval);
-                    c.setRetryConfiguration(retry);
-                });
+                c.setRetryConfiguration(RetryConfigurationResolver.resolve(extConfig.retry));
             }
         });
